fastapi_swagger/requirements.txt
exa-py==1.0.1
openai==1.12.0
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.5.2
supabase-py>=2.0.0
reportlab==4.3.1
httpx==0.26.0
